Fix radionuclide source normalization
Fix a bug where particles were killed by setting their energy to zero,
but not their weight, during source generation in
EGS_RadionuclideSource. This resulted in extra dose being deposited
since the rest mass is added to the source particle energy in shower().

Fix a bug where the branching ratio was not used for alpha records in
EGS_RadionuclideSource. Add a check to make sure the branching ratios
add up to 1 for the products of a given parent nuclide. If not, the
missing branch is assumed to be fission. Add support for P and Q shells
for the initial vacancies of relaxations from decays.

Change the relaxation emissions to use the intensity of each shell,
instead of the total intensity from the ENSDF file. This was necessary
since the shell intensities did not always add up to the total
intensity.

Improve compatibility with NNDC ensdf files, by allowing for the
normalization to be neglected. Also improve the robustness of checking
comments for atomic relaxation data.